fix the id name map
